+++ packages/components/src/components/primitives/form/DropdownField/DropdownNativeSelect/DropdownNativeSelect.tsx
@@ -3,27 +3,9 @@ import cn from 'classnames';
 import { Icon } from 'components/primitives';
 import { expand_more as ExpandMore } from 'components/primitives/Icon/icons/navigation';
 import { getDropdownSingleSelectClasses } from 'utils/getDropdownSingleSelectClasses';
+import { SelectProps, DROPDOWN_DEFAULT_TEST_ID } from '../DropdownSingleSelect';
 
-export const DROPDOWN_NATIVE_DEFAULT_TEST_ID = 'dropdown_native';
-
-export interface Option {
-  label: any;
-  value: any;
-}
-
-type Props = {
-  className?: string;
-  options?: Option[];
-  placeHolder?: string;
-  initialSelectedOption?: Option;
-  onItemChange?: (item: Option) => void;
-  testId?: string;
-  disabled?: boolean;
-  name?: string;
-  invalid?: boolean;
-};
-
-export const DropdownNativeSelect = forwardRef<HTMLSelectElement, Props>(
+export const DropdownNativeSelect = forwardRef<HTMLSelectElement, SelectProps>(
   (
     {
       className,
@@ -31,14 +13,14 @@ export const DropdownNativeSelect = forwardRef<HTMLSelectElement, Props>(
       placeHolder = 'Select an item',
       initialSelectedOption,
       onItemChange,
-      testId = DROPDOWN_NATIVE_DEFAULT_TEST_ID,
+      testId = DROPDOWN_DEFAULT_TEST_ID,
       disabled = false,
       name = 'dropdown',
       invalid,
-    }: Props,
+    }: SelectProps,
     ref
   ) => {
-    const defaultValue = initialSelectedOption && initialSelectedOption.value;
+    const defaultValue = initialSelectedOption?.value;
     const [itemSelected, setItemSelected] = useState(defaultValue || '');
 
     const {
@@ -54,15 +36,10 @@ export const DropdownNativeSelect = forwardRef<HTMLSelectElement, Props>(
 
     const handleOnChange = (event) => {
       const index = event.target.selectedIndex;
-      const label = event.target[index].label;
-      event.target.className += ' !text-fill-forms-filled';
+      const value = event.target.value;
 
-      setItemSelected(event.target.value);
-
-      return onItemChange({
-        label: label,
-        value: event.target.value,
-      });
+      setItemSelected(value);
+      return onItemChange(options[index]);
     };
 
     return (
@@ -85,13 +62,13 @@ export const DropdownNativeSelect = forwardRef<HTMLSelectElement, Props>(
         >
           <option value="">{placeHolder}</option>
           {options.map((item, index) => (
-            <option key={`${index}${item.value}`} value={item.value}>
+            <option key={index} value={item.value}>
               {item.label}
             </option>
           ))}
         </select>
         <div className={iconWrapper}>
-          <Icon width={24} height={24} className={icon}>
+          <Icon className={icon}>
             <ExpandMore type="filled" />
           </Icon>
         </div>
